add near circle like function

+++ app/src/main/java/com/hwl/beta/mq/bean/ChatFriendRequestBean.java
@@ -68,7 +68,7 @@ public class ChatFriendRequestBean {
         return toUserId;
     }
 
-    public void setToUserId(int toUserId) {
+    public void setToUserId(long toUserId) {
         this.toUserId = toUserId;
     }
